wifi: Change returned type for CtrlBAckResponseHeader::IndexInBitmap

+++ src/wifi/model/ctrl-headers.h
@@ -374,7 +374,7 @@ private:
    * to set to 1 in the compressed bitmap to indicate that packet having
    * sequence number equals to <i>seq</i> was correctly received.
    */
-  uint8_t IndexInBitmap (uint16_t seq) const;
+  uint16_t IndexInBitmap (uint16_t seq) const;
 
   /**
    * Checks if sequence number <i>seq</i> can be acknowledged in the bitmap.
